feat: replace kill command

+++ src/main/kotlin/model/command/factory/KillCommandFactory.kt
@@ -3,11 +3,11 @@ package model.command.factory
 import common.OsType
 
 interface KillCommandFactory {
-    fun create(pid: Int): List<String>
+    fun create(pid: Long): List<String>
 
     private class KillCommandFactoryForMac : KillCommandFactory {
         @OptIn(kotlin.ExperimentalStdlibApi::class)
-        override fun create(pid: Int): List<String> {
+        override fun create(pid: Long): List<String> {
             return buildList {
                 add("kill")
                 add("-SIGINT")
@@ -18,7 +18,7 @@ interface KillCommandFactory {
 
     private class KillCommandFactoryForWindows : KillCommandFactory {
         @OptIn(kotlin.ExperimentalStdlibApi::class)
-        override fun create(pid: Int): List<String> {
+        override fun create(pid: Long): List<String> {
             return buildList {
                 add("taskkill")
                 add("/PID")
